getIdEtiqueta
Las funcioens en useEtiqueta no usan mas useState, es innecesario, asi que simplemente retornan el resultado. Ahora TareasEtiquetas se ahorra la llamada al back gracias a la nueva funcion.

+++ gdtms_frontend/src/componentes/TareasCompletas.js
@@ -34,14 +34,18 @@ export function TareasCompletas(){
     //CONTEXTO: ME TIRA ERROR
 
     const limpiarTareas = ()=>{
-        axios.delete(`http://localhost:3001/tareas/completas?userId=${userId}`)
-        .then((res)=>{
-            if (res) {
-                console.log("Backend: ", res.data);
-                actualizarTareas();
-            }
-            else console.log("No hubo respuesta limpiarTareas() desde el backend");
-        }).catch((err)=> console.log("Ha ocurrido un error en limpiarTareas(): ", err))
+        let confirmar = window.confirm("Estas seguro de eliminar esta tarea?");
+        if(confirmar === true){
+            axios.delete(`http://localhost:3001/tareas/completas?userId=${userId}`)
+            .then((res)=>{
+                if (res) {
+                    console.log("Backend: ", res.data);
+                    actualizarTareas();
+                }
+                else console.log("No hubo respuesta limpiarTareas() desde el backend");
+            }).catch((err)=> console.log("Ha ocurrido un error en limpiarTareas(): ", err))
+        }
+        else return
     }
 
     useEffect(() => {
